tgbot::handler
Struct WebhookServer

pub struct WebhookServer { /* private fields */ }

Available on crate feature webhook only.

Represents a simple webhook server for handling incoming updates from the Telegram Bot API.

Implementations

impl WebhookServer

pub fn new<A, B>(path: A, handler: B) -> Self
where A: AsRef<str>, B: UpdateHandler + Send + Sync + 'static,

Creates a new WebhookServer.

Arguments

path - The path where the webhook server will receive incoming updates.

handler - The handler for processing updates.

pub async fn run<T>(self, address: T) -> Result<SocketAddr, IoError>
where T: Into<SocketAddr>,

Runs the server

Returns the local address that the server is bound to.

Arguments

address - The address to bind the server to.