Module tgbot::methods

Methods available in the Bot API

Structs

AddStickerToSet
Add a new sticker to a set created by the bot

AnswerCallbackQuery
Send answer to callback query sent from inline keyboard

AnswerInlineQuery
Use this method to send answers to an inline query

AnswerPreCheckoutQuery
Respond to pre-checkout query

AnswerShippingQuery
Reply to shipping query

CreateNewStickerSet
Create new sticker set owned by a user

DeleteChatPhoto
Delete a chat photo

DeleteChatStickerSet
Delete a group sticker set from a supergroup

DeleteMessage
Delete a message, including service messages

DeleteStickerFromSet
Delete a sticker from a set created by the bot

DeleteWebhook
Remove webhook integration if you decide to switch back to getUpdates

EditMessageCaption
Edit caption of message sent by the bot or via the bot (for inline bots)

EditMessageLiveLocation
Edit live location messages sent by the bot or via the bot (for inline bots)

EditMessageMedia
Edit audio, document, photo, or video messages

EditMessageReplyMarkup
Edit only the reply markup of messages sent by the bot or via the bot (for inline bots)

EditMessageText
Edit text and game messages sent by the bot or via the bot (for inline bots)

ExportChatInviteLink
Generate a new invite link for a chat

ForwardMessage
Forward message of any kind

GetChat
Get up to date information about the chat

GetChatAdministrators
Get a list of administrators in a chat

GetChatMember
Get information about a member of a chat

GetChatMembersCount
Get the number of members in a chat

GetFile
Get basic info about a file and prepare it for downloading

GetGameHighScores
Get data for high score tables

GetMe
Returns basic information about the bot

GetStickerSet
Get a sticker set

GetUpdates
Receive incoming updates using long polling

GetUserProfilePhotos
Get a list of profile pictures for a user

GetWebhookInfo
Get current webhook status

KickChatMember
Kick a user from a group, a supergroup or a channel

LeaveChat
Leave a group, supergroup or channel

PinChatMessage
Pin a message in a group, supergroup or a channel

PromoteChatMember
Promote or demote a user in a supergroup or a channel

RestrictChatMember
Restrict a user in a supergroup

SendAnimation
Send animation files (GIF or H.264/MPEG-4 AVC video without sound)

SendAudio
Send audio files

SendChatAction
Tell the user that something is happening on the bot's side

SendContact
Send phone contacts

SendDocument
Send general files

SendGame
Use this method to send a game

SendInvoice
Send invoice

SendLocation
Send point on the map

SendMediaGroup
Send a group of photos or videos as an album

SendMessage
Send text messages

SendPhoto
Send photo

SendPoll
Use this method to send a native poll

SendQuiz
Use this method to send a quiz

SendSticker
Send .webp sticker

SendVenue
Send information about a venue

SendVideo
Send video file

SendVideoNote
Send video message

SendVoice
Send audio files, if you want Telegram clients to display the file as a playable voice message

SetChatAdministratorCustomTitle
Set a custom title for an administrator in a supergroup promoted by the bot

SetChatDescription
Change the description of a group, a supergroup or a channel

SetChatPermissions
Set default chat permissions for all members

SetChatPhoto
Set a new profile photo for the chat

SetChatStickerSet
Set a new group sticker set for a supergroup

SetChatTitle
Change the title of a chat

SetGameScore
Set the score of the specified user in a game

SetPassportDataErrors
Informs a user that some of the Telegram Passport elements they provided contains errors

SetStickerPositionInSet
Move a sticker in a set created by the bot to a specific position

SetWebhook
Specify a url and receive incoming updates via an outgoing webhook

StopMessageLiveLocation
Stop updating a live location message sent by the bot or via the bot (for inline bots) before live_period expires

StopPoll
Use this method to stop a poll which was sent by the bot

UnbanChatMember
Unban a previously kicked user in a supergroup or channel

UnpinChatMessage
Unpin a message in a group, supergroup or a channel

UploadStickerFile
Upload a .png file with a sticker for later use in createNewStickerSet and addStickerToSet methods

Traits

Method
Represents an API method

Type Definitions

StopQuiz
Use this method to stop a quiz which was sent by the bot