Struct tgbot::methods::PromoteChatMember

pub struct PromoteChatMember { /* fields omitted */ }

Promote or demote a user in a supergroup or a channel

The bot must be an administrator in the chat for this to work and must have the appropriate admin rights Pass False for all boolean parameters to demote a user

Methods

impl PromoteChatMember

pub fn new<C: Into<ChatId>>(chat_id: C, user_id: Integer) -> Self

Creates a new PromoteChatMember with empty optional parameters

Arguments

chat_id - Unique identifier for the target chat

user_id - Unique identifier of the target user

pub fn promote_all(self) -> Self

Promote all privileges

pub fn demote_all(self) -> Self

Demote all privileges

pub fn can_change_info(self, can_change_info: bool) -> Self

Administrator can change chat title, photo and other settings

pub fn can_post_messages(self, can_post_messages: bool) -> Self

Administrator can create channel posts, channels only

pub fn can_edit_messages(self, can_edit_messages: bool) -> Self

Administrator can edit messages of other users and can pin messages, channels only

pub fn can_delete_messages(self, can_delete_messages: bool) -> Self

Administrator can delete messages of other users

pub fn can_invite_users(self, can_invite_users: bool) -> Self

Administrator can invite new users to the chat

pub fn can_restrict_members(self, can_restrict_members: bool) -> Self

Administrator can restrict, ban or unban chat members

pub fn can_pin_messages(self, can_pin_messages: bool) -> Self

Administrator can pin messages, supergroups only

pub fn can_promote_members(self, can_promote_members: bool) -> Self

Administrator can add new administrators with a subset of his own privileges or demote administrators that he has promoted, directly or indirectly (promoted by administrators that were appointed by him)