Module tgbot::types

Telegram Bot API types

Re-exports

pub use mime;

Structs

AddStickerToSet

Adds a new sticker to a set created by the bot.

Animation

Represents an animation file (GIF or H.264/MPEG-4 AVC video without sound).

AnswerCallbackQuery

Sends an answer to a callback query sent from an inline keyboard.

AnswerInlineQuery

Sends an answer to an inline query.

AnswerPreCheckoutQuery

Responds to a pre-checkout query.

AnswerShippingQuery

Replies to a shipping query.

AnswerWebAppQuery

Sets a result of an interaction with a Web App and send a corresponding message on behalf of the user to the chat from which the query originated

ApproveChatJoinRequest

Approves a chat join request.

Audio

Represents an audio file to be treated as music by the Telegram clients.

BanChatMember

Bans a user in a chat.

BanChatSenderChat

Bans a channel chat in a supergroup or a channel.

Birthdate

Represents the date of birth of a user.

Bot

Represents information about a bot returned in GetBot.

BotCommand

Represents a command of a bot.

BotDescription

Represents a description of a bot.

BotName

Represents a name of a bot.

BotShortDescription

Represents a short description of a bot.

BusinessConnection

Describes the connection of the bot with a business account.

BusinessIntro

Represents the intro of the business.

BusinessLocation

Provides information about address and location of the business.

BusinessMessagesDeleted

Provides information about messages deleted from a connected business account.

BusinessOpeningHours

Provides information about the opening hours of the business.

BusinessOpeningHoursInterval

Provides information about the time interval describing business opening hours.

CallbackQuery

Represents an incoming callback query from a callback button in an inline keyboard.

ChannelChat

Represents a channel chat.

ChatAdministratorRights

Represents the rights of an administrator in a chat.

ChatBackground

Represents a chat background.

ChatBoost

Contains information about a chat boost.

ChatBoostRemoved

Represents a boost removed from a chat.

ChatBoostSourceGiveaway

The boost was obtained by the creation of a Telegram Premium giveaway.

ChatBoostUpdated

Represents a boost added to a chat or changed.

ChatFullInfo

Contains full information about a chat.

ChatInviteLink

Represents an invite link for a chat.

ChatJoinRequest

Represents a join request sent to a chat.

ChatLocation

Represents a location to which a chat is connected.

ChatMemberAdministrator

Represents a chat administrator.

ChatMemberCreator

Represents a chat member that owns the chat and has all administrator privileges.

ChatMemberKicked

Represents a kicked chat member.

ChatMemberRestricted

Represents a restricted user.

ChatMemberUpdated

Represents changes in a status of a chat member.

ChatPeerId

ID of a chat.

ChatPermissions

Represents actions that a non-administrator user is allowed to take in a chat.

ChatPhoto

Represents a chat photo.

ChatUsername

Username of a chat in the format @username.

ChosenInlineResult

Represents a result of an inline query that was chosen by the user and sent to their chat partner.

Close

Closes a bot instance before moving it from one local server to another.

CloseForumTopic

Closes an open topic in a forum supergroup chat.

CloseGeneralForumTopic

Closes an opened ‘General’ topic in a forum supergroup chat.

Command

A simple command implementation.

Contact

Represents a phone contact.

CopyMessage

Copies a message.

CopyMessages

Copies messages of any kind.

CreateChatInviteLink

Creates an additional invite link for a chat.

CreateForumTopic

Creates a topic in a forum supergroup chat.

CreateInvoiceLink

Creates a link for an invoice.

CreateNewStickerSet

Creates a new sticker set owned by a user.

DeclineChatJoinRequest

Declines a chat join request.

DeleteBotCommands

Deletes a list of a bot commands of a given scope and user language.

DeleteChatPhoto

Deletes a chat photo.

DeleteChatStickerSet

Deletes a group sticker set from a supergroup.

DeleteForumTopic

Deletes a forum topic along with all its messages in a forum supergroup chat.

DeleteMessage

Deletes a message.

DeleteMessages

Deletes multiple messages simultaneously.

DeleteStickerFromSet

Deletes a sticker from a set created by the bot.

DeleteStickerSet

Deletes a sticker set that was created by the bot.

DeleteWebhook

Removes a webhook integration if you decide to switch back to crate::types::GetUpdates.

Dice

Represents a dice with a random value.

Document

Represents a general file (as opposed to photos, voice messages and audio files).

EditChatInviteLink

Changes a non-primary invite link created by a bot.

EditForumTopic

Changes name and icon of a topic in a forum supergroup chat.

EditGeneralForumTopic

Changes the name of the ‘General’ topic in a forum supergroup chat.

EditMessageCaption

Changes a caption of a message.

EditMessageLiveLocation

Changes a live location message.

EditMessageMedia

Changes animation, audio, document, photo, or video message.

EditMessageReplyMarkup

Changes the reply markup of a message.

EditMessageText

Changes a text or a game message.

EncryptedCredentials

Represents a data required for decrypting and authenticating EncryptedPassportElement.

EncryptedPassportElementAddress

Represents an address.

EncryptedPassportElementBankStatement

Represents a bank statement.

EncryptedPassportElementDriverLicense

Represents a driver license.

EncryptedPassportElementEmail

Represents an E-Mail.

EncryptedPassportElementIdentityCard

Represents an identity card.

EncryptedPassportElementInternalPassport

Represents an internal passport.

EncryptedPassportElementPassport

Represents a passport.

EncryptedPassportElementPassportRegistration

Represents a passport registration.

EncryptedPassportElementPersonalDetails

Represents personal details.

EncryptedPassportElementPhoneNumber

Represents a phone number.

EncryptedPassportElementRentalAgreement

Represents a rental agreement.

EncryptedPassportElementTemporaryRegistration

Represents a temporary registration.

EncryptedPassportElementUtilityBill

Represents an utility bill.

ExportChatInviteLink

Generates a new invite link for a chat.

ExternalReplyInfo

Contains information about a message that is being replied to, which may come from another chat or forum topic.

File

Represents a file ready to be downloaded.

ForceReply

Represents a reply interface trigger.

ForumTopic

Represents a forum topic.

ForwardMessage

Forwards a message.

ForwardMessages

Forwards multiple messages.

Game

Represents a Game.

GameHighScore

Represents a row of the high scores table for a game.

GetBot

Returns a basic information about a bot.

GetBotCommands

Returns the current list of bot commands.

GetBotDefaultAdministratorRights

Returns the current default administrator rights of a bot.

GetBotDescription

Returns the current description of a bot for a given user language.

GetBotName

Returns the current name of a bot for a given user language.

GetBotShortDescription

Returns the current short description of a bot for a given user language.

GetBusinessConnection

Returns information about the connection of the bot with a business account.

GetChat

Returns up to date information about the chat.

GetChatAdministrators

Returns a list of administrators in a chat.

GetChatMember

Returns a member of a chat.

GetChatMemberCount

Returns a number of members in a chat.

GetChatMenuButton

Returns the current value of the bot menu button in a private chat, or the default menu button.

GetCustomEmojiStickers

Returns information about custom emoji stickers by their identifiers.

GetFile

Returns a basic information about a file and prepares it for downloading.

GetForumTopicIconStickers

Returns custom emoji stickers, which can be used as a forum topic icon by any user.

GetGameHighScores

Returns data for high score tables.

GetStarTransactions

Returns the bot’s Telegram Star transactions in chronological order.

GetStickerSet

Returns a sticker set.

GetUpdates

Returns incoming updates using long polling.

GetUserChatBoosts

Returns the list of boosts added to a chat by a user.

GetUserProfilePhotos

Returns a list of profile pictures for a user.

GetWebhookInfo

Returns current webhook status.

Giveaway

Represents a message about a scheduled giveaway.

GiveawayCompleted

Represents a service message about the completion of a giveaway without public winners.

GiveawayWinners

Represents a message about the completion of a giveaway with public winners.

GroupChat

Represents a group chat.

HideGeneralForumTopic

Hides the ‘General’ topic in a forum supergroup chat.

InaccessibleMessage

Describes a message that was deleted or is otherwise inaccessible to the bot.

InlineKeyboardButton

Represents a button of an inline keyboard.

InlineKeyboardMarkup

Represents an inline keyboard that appears right next to the message it belongs to.

InlineQuery

Represents an incoming inline query.

InlineQueryResultArticle

Represents a link to an article or a web page.

InlineQueryResultAudio

Represents a link to an mp3 audio file.

InlineQueryResultCachedAudio

Represents a link to an mp3 audio file stored on the Telegram servers.

InlineQueryResultCachedDocument

Represents a link to a file stored on the Telegram servers.

InlineQueryResultCachedGif

Link to an animated GIF file stored on the Telegram servers.

InlineQueryResultCachedMpeg4Gif

Represents a link to a video animation (H.264/MPEG-4 AVC video without sound) stored on the Telegram servers.

InlineQueryResultCachedPhoto

Represents a link to a photo stored on the Telegram servers.

InlineQueryResultCachedSticker

Represents a link to a sticker stored on the Telegram servers.

InlineQueryResultCachedVideo

Represents a link to a video file stored on the Telegram servers.

InlineQueryResultCachedVoice

Represents a link to a voice message stored on the Telegram servers.

InlineQueryResultContact

Represents a contact with a phone number.

InlineQueryResultDocument

Represents a link to a file.

InlineQueryResultGame

Represents a game.

InlineQueryResultGif

Represents a link to an animated GIF file.

InlineQueryResultLocation

Represents a location on a map.

InlineQueryResultMpeg4Gif

Represents a link to a video animation (H.264/MPEG-4 AVC video without sound).

InlineQueryResultPhoto

Represents a link to a photo.

InlineQueryResultVenue

Represents a venue.

InlineQueryResultVideo

Represents a link to a page containing an embedded video player or a video file.

InlineQueryResultVoice

Represents a link to a voice recording in an OGG container encoded with OPUS.

InlineQueryResultsButton

Represents a button to be shown above inline query results.

InputFileReader

Represents a file reader for uploading files.

InputMedia

Represents a content of a media message to be sent.

InputMediaAnimation

Represents an animation file (GIF or H.264/MPEG-4 AVC video without sound) to be sent.

InputMediaAudio

Represents an audio file to be treated as music to be sent.

InputMediaDocument

Represents a general file to be sent.

InputMediaPhoto

Represents a photo to be sent.

InputMediaVideo

Represents a video to be sent.

InputMessageContentContact

Represents a contact message to be sent as the result of an inline query.

InputMessageContentInvoice

Represents an invoice message to be sent as the result of an inline query.

InputMessageContentLocation

Represents a location message to be sent as the result of an inline query.

InputMessageContentText

Represents a text message to be sent as the result of an inline query.

InputMessageContentVenue

Represents a venue message to be sent as the result of an inline query.

InputPaidMediaGroup

Describes the paid media group to be sent.

InputPaidMediaGroupItem

Represents an input paid media.

InputPaidMediaVideo

The paid media to send is a video.

InputPollOption

Contains information about one answer option in a poll to send.

InputSticker

Represents a metadata of a sticker to be added to a sticker set.

InputStickers

Represents a collection of stickers to be added to a sticker set.

Invoice

Represents an invoice.

InvoiceParameters

Represents an invoice parameters used in CreateInvoiceLink and SendInvoice.

KeyboardButton

Represents a button of the reply keyboard.

KeyboardButtonPollType

Represents a type of a poll which is allowed to be created and sent when the corresponding button is pressed.

KeyboardButtonRequestChat

Represents a criteria used to request a suitable chat.

KeyboardButtonRequestUsers

Represents a criteria used to request a suitable user.

LabeledPrice

Represents a portion of the price for goods or services.

LeaveChat

Leaves a group, supergroup or channel.

LinkPreviewOptions

Represents the options used for link preview generation.

Location

Represents a point on a map.

LogOut

Logs out from the Cloud Bot API.

LoginUrl

Represents a parameter of the inline keyboard button used to automatically authorize a user.

MaskPosition

Represents a position on faces where a mask should be placed by default.

MediaGroup

Represents a group of input media to be sent.

MediaGroupItem

Represents a media group item.

MenuButtonWebApp

Represents a menu button, which launches a Web App.

Message

Represents a message.

MessageDataAudio

Represents an audio message data.

MessageDataAutoDeleteTimer

Represents a service message about a change in auto-delete timer settings.

MessageDataChatShared

Represents an information about the chat whose identifier was shared with the bot using a crate::types::KeyboardButtonRequestChat button.

MessageDataDocument

Represents an document message data.

MessageDataForumTopicCreated

Represents a service message about a new forum topic created in the chat.

MessageDataForumTopicEdited

Represents a service message about an edited forum topic.

MessageDataPhoto

Represents a list of available sizes of the photo.

MessageDataProximityAlert

Represents a content of a service message, sent whenever a user in the chat triggers a proximity alert set by another user.

MessageDataUsersShared

Contains information about the users whose identifiers were shared with the bot using a crate::types::KeyboardButton::with_request_users button.

MessageDataVideo

Represents a video message data.

MessageDataVideoChatEnded

Represents a service message about a video chat ended in the chat.

MessageDataVideoChatParticipantsInvited

A service message about new members invited to a video chat.

MessageDataVideoChatScheduled

A service message about a video chat scheduled in the chat.

MessageDataVoice

Message is a voice message, information about the file.

MessageDataWriteAccess

Represents a service message about a user allowing a bot to write messages after adding it to the attachment menu, launching a Web App from a link, or accepting an explicit request from a Web App sent by the method requestWriteAccess.

MessageId

Represents an unique message identifier.

MessageOriginChannel

The message was originally sent to a channel chat.

MessageOriginChat

The message was originally sent on behalf of a chat to a group chat.

MessageOriginHiddenUser

The message was originally sent by an unknown user.

MessageOriginUser

The message was originally sent by a known user.

MessageReactionCountUpdated

Represents reaction changes on a message with anonymous reactions.

MessageReactionUpdated

Represents a change of a reaction on a message performed by a user.

OrderInfo

Represents an order.

PaidMediaInfo

Describes the paid media added to a message.

PaidMediaPreview

The paid media isn’t available before the payment.

PassportData

Represents a telegram Passport data shared with a bot by a user.

PassportElementError

Represents an error in the Telegram Passport element which was submitted that should be resolved by a user.

PassportFile

Represents a file uploaded to Telegram Passport.

PhotoSize

Represents a size of a photo or a file / sticker thumbnail.

PinChatMessage

Adds a message to a list of pinned messages in a chat.

PollAnswer

Represents an answer of a user in a non-anonymous poll.

PollOption

Represents an answer option in a poll.

PreCheckoutQuery

Represents an incoming pre-checkout query.

PrivateChat

Represents a private chat.

ProfileAccentColor

Represents a profile accent color.

PromoteChatMember

Promotes or demotes a user in a chat.

Quiz

Represents a quiz.

ReactionCount

Represents a reaction added to a message along with the number of times it was added.

RefundStarPayment

Refunds a successful payment in Telegram Stars.

RegularPoll

Represents a regular poll.

ReopenForumTopic

Reopens a closed topic in a forum supergroup chat.

ReopenGeneralForumTopic

Reopens a closed ‘General’ topic in a forum supergroup chat.

ReplaceStickerInSet

Replaces an existing sticker in a sticker set with a new one.

ReplyKeyboardMarkup

Represents a custom keyboard with reply options.

ReplyKeyboardRemove

Represents a trigger to remove the custom keyboard.

ReplyParameters

Describes reply parameters for the message that is being sent.

ReplyQuote

Quoted part of the message to be replied to.

ResponseError

Represents a response error.

RestrictChatMember

Restricts a user in a supergroup.

RevokeChatInviteLink

Revokes an invite link created by a bot.

SendAnimation

Sends an animation file (GIF or H.264/MPEG-4 AVC video without sound).

SendAudio

Sends am audio file.

SendChatAction

Tells a user that something is happening on a bot side.

SendContact

Sends a phone contact.

SendDice

Sends a dice.

SendDocument

Sends a general file.

SendGame

Sends a game.

SendInvoice

Sends an invoice.

SendLocation

Sends a point on a map.

SendMediaGroup

Sends a group of photos or videos as an album.

SendMessage

Sends a text message.

SendPaidMedia

Send paid media to channel chats.

SendPhoto

Sends a photo.

SendPoll

Sends a native poll.

SendQuiz

Sends a quiz.

SendSticker

Sends a static WEBP, animated TGS, or video WEBM sticker.

SendVenue

Sends information about a venue.

SendVideo

Sends a video file.

SendVideoNote

Sends a video message.

SendVoice

Sends a voice message.

SentWebAppMessage

Represents an inline message sent by a Web App on behalf of a user

SetBotCommands

Changes a list of commands of a bot.

SetBotDefaultAdministratorRights

Changes default administrator rights requested by a bot when it’s added as an administrator to groups or channels.

SetBotDescription

Changes the description of a bot, which is shown in a chat with the bot if the chat is empty.

SetBotName

Changes the name of a bot.

SetBotShortDescription

Changes the short description of a bot, which is shown on the bot profile page and is sent together with the link when users share the bot.

SetChatAdministratorCustomTitle

Sets a custom title for an administrator in a supergroup promoted by the bot.

SetChatDescription

Changes a description of a chat.

SetChatMenuButton

Changes a button of a menu of a bot in a private chat, or a default menu button.

SetChatPermissions

Sets default chat permissions for all members.

SetChatPhoto

Sets a new profile photo for a chat.

SetChatStickerSet

Sets a new group sticker set for a supergroup.

SetChatTitle

Changes a title of a chat.

SetCustomEmojiStickerSetThumbnail

Sets the thumbnail of a custom emoji sticker set.

SetGameScore

Sets the score of the specified user in a game.

SetMessageReaction

Changes the chosen reactions on a message.

SetPassportDataErrors

Informs a user that some of the Telegram Passport elements they provided contains errors.

SetStickerEmojiList

Changes the list of emoji assigned to a regular or custom emoji sticker.

SetStickerKeywords

Changes search keywords assigned to a regular or custom emoji sticker.

SetStickerMaskPosition

Changes the mask position of a mask sticker.

SetStickerPositionInSet

Moves a sticker in a set created by the bot to a specific position.

SetStickerSetThumbnail

Sets a thumbnail of a sticker set.

SetStickerSetTitle

Sets a title of a created sticker set.

SetWebhook

Specifies an url and returns incoming updates via an outgoing webhook.

SharedUser

Contains information about a user that was shared with the bot using a crate::types::KeyboardButtonRequestUsers button.

ShippingAddress

Represents a shipping address.

ShippingOption

Represents a shipping option.

ShippingQuery

Represents a shipping query.

StarTransaction

Describes a Telegram Star transaction.

StarTransactions

Contains a list of Telegram Star transactions.

Sticker

Represents a sticker.

StickerSet

Represents a sticker set.

StopMessageLiveLocation

Stops updating a live location message before live_period expires.

StopPoll

Stops a poll which was sent by the bot.

Story

Represents a story.

SuccessfulPayment

Represents a successful payment.

SupergroupChat

Represents a supergroup chat.

SwitchInlineQueryChosenChat

Represents an inline button that switches the current user to inline mode in a chosen chat, with an optional default inline query.

Text

Represents a text with entities.

TextEntities

Represents a collection of text entities.

TextEntityBotCommand

Represents a bot command found in text.

TextEntityPosition

Represents a position of an entity in a text.

TextQuote

Contains information about the quoted part of a message that is replied to by the given message.

UnbanChatMember

Unbans a previously kicked user in a supergroup or channel.

UnbanChatSenderChat

Unbans a previously banned channel chat in a supergroup or channel.

UnexpectedEncryptedPassportElementType

Represents an error with unexpected encrypted passport element type.

UnexpectedUpdate

Conversion of an Update into T failed.

UnhideGeneralForumTopic

Reveals the ‘General’ topic in a forum supergroup chat.

UnpinAllChatMessages

Clears a list of pinned messages in a chat.

UnpinAllForumTopicMessages

Clears the list of pinned messages in a forum topic.

UnpinAllGeneralForumTopicMessages

Clears the list of pinned messages in a General forum topic.

UnpinChatMessage

Removes a message from a list of pinned messages in a chat.

Update

Represents an incoming update.

UploadStickerFile

Uploads a file with a sticker for later use in the CreateNewStickerSet and AddStickerToSet methods.

User

Represents a user.

UserChatBoosts

Represents a list of boosts added to a chat by a user.

UserPeerId

ID of a user.

UserProfilePhotos

Represents a list of profile pictures of a user.

UserUsername

Username of a user in the format @username.

Venue

Represents a venue.

Video

Represents a video file.

VideoNote

Represents a video message.

Voice

Represents a voice file.

WebAppData

Represents a data sent from a Web App to the bot.

WebAppInfo

Represents a Web App.

WebhookInfo

Represents a current status of a webhook.

Enums

AccentColor

Represents an accent color.

AccentColorError

Represents an accent color error.

AllowedUpdate

Represents a type of update to receive.

BackgroundFill

Describes the way a background is filled based on the selected colors.

BackgroundType

This object describes the type of a background.

BotCommandError

Represents an error that can occur when creating a new BotCommand.

BotCommandScope

Represents a scope to which bot commands are applied.

CallbackQueryError

Represents an error that can occur while parsing data from a callback query.

Chat

Represents a chat.

ChatAction

Represents a type of an action to tell a user that something is happening on a bot side.

ChatBoostSource

Describes the source of a chat boost.

ChatFullInfoType

Type of the chat.

ChatId

Represents an ID or username of a chat.

ChatMember

Represents a member of a chat.

CommandError

Represents an error when parsing a command.

DiceType

Represents a type of a dice.

EditMessageResult

Represents a result of EditMessage* requests.

EncryptedPassportElement

Represents an information about documents or other Telegram Passport elements shared with a bot by a user.

EncryptedPassportElementType

Represents a type of an encrypted passport element.

ExternalReplyData

Contains data of an external reply info.

ForumTopicIconColor

Represents a color of a forum topic icon.

InlineKeyboardButtonType

Represents a type of an inline keyboard button.

InlineKeyboardError

Represents an error occurred with an inline keyboard.

InlineQueryChatType

Represents a type of the chat, from which the inline query was sent.

InlineQueryResult

Represents a result of an inline query.

InputFile

Represents a file to upload.

InputMediaError

An error occurred with InputMedia.

InputMediaType

Represents a metadata of the input media.

InputMessageContent

Represents a content of a message to be sent as a result of an inline query.

InputPaidMediaGroupError

Describes an crate::types::InputPaidMediaGroup error.

InputStickerError

Represents an error when converting input stickers into a multipart form.

MaskPositionPoint

Represents a part of the face relative to which the mask should be placed.

MaybeInaccessibleMessage

Describes a message that can be inaccessible to the bot.

MediaGroupError

Represents a media group error.

MentionError

Represents an error occurred when getting user mention.

MenuButton

Represents a menu button of the bot in a private chat.

MessageData

Represents a message data.

MessageOrigin

Describes the origin of a message.

MessageSender

Represents a sender of a message.

PaidMedia

Describes paid media.

ParseMode

Represents a mode for parsing entities in a text.

Poll

Represents a poll.

PollAnswerVoter

Represents the chat or the user that changed answer to the poll.

PollType

Represents a type of a poll.

ProfileAccentColorError

Represents a profile accent color error.

ReactionType

Represents a reaction type.

ReplyMarkup

Represents a reply markup.

ReplyMarkupError

Represents an error occurred with reply markup.

ReplyParametersError

Represents an error occurred with reply markup.

ReplyTo

Contains information about a message or a story that is being replied to.

Response

Represents an API Response.

RevenueWithdrawalState

Describes the state of a revenue withdrawal operation.

SendAnimationError

Represents an error when sending an animation.

SendAudioError

Represents an error when sending an audio.

SendDocumentError

Represents an error when sending a document.

SendVideoError

Represents an error when sending a video.

SendVideoNoteError

Represents an error when sending a video note.

StickerFormat

Represents a format of stickers in the set.

StickerType

Represents a type of stickers in the set.

TextEntity

Represents an entity in a text.

TextEntityError

Represents an error when parsing/serializing entities.

TransactionPartner

Describes the source of a transaction, or its recipient for outgoing transactions.

UpdateType

Represents a type of an update.

UserId

Represents a user ID.

Type Aliases

Float

Represents a Telegram Float type

Integer

Represents a Telegram Integer type

StopQuiz

Stops a quiz which was sent by the bot.