Enum tgbot::types::InlineQueryResult

pub enum InlineQueryResult {

    Article(InlineQueryResultArticle),
    Audio(InlineQueryResultAudio),
    CachedAudio(InlineQueryResultCachedAudio),
    CachedDocument(InlineQueryResultCachedDocument),
    CachedGif(InlineQueryResultCachedGif),
    CachedMpeg4Gif(InlineQueryResultCachedMpeg4Gif),
    CachedPhoto(InlineQueryResultCachedPhoto),
    CachedSticker(InlineQueryResultCachedSticker),
    CachedVideo(InlineQueryResultCachedVideo),
    CachedVoice(InlineQueryResultCachedVoice),
    Contact(InlineQueryResultContact),
    Document(InlineQueryResultDocument),
    Game(InlineQueryResultGame),
    Gif(InlineQueryResultGif),
    Location(InlineQueryResultLocation),
    Mpeg4Gif(InlineQueryResultMpeg4Gif),
    Photo(InlineQueryResultPhoto),
    Venue(InlineQueryResultVenue),
    Video(InlineQueryResultVideo),
    Voice(InlineQueryResultVoice),

}

Result of an inline query

Variants

Article(InlineQueryResultArticle)

Link to an article or web page

Tuple Fields of Article

0: InlineQueryResultArticle

Audio(InlineQueryResultAudio)

Link to an mp3 audio file

Tuple Fields of Audio

0: InlineQueryResultAudio

CachedAudio(InlineQueryResultCachedAudio)

Link to an mp3 audio file stored on the Telegram servers

Tuple Fields of CachedAudio

0: InlineQueryResultCachedAudio

CachedDocument(InlineQueryResultCachedDocument)

Link to a file stored on the Telegram servers

Tuple Fields of CachedDocument

0: InlineQueryResultCachedDocument

CachedGif(InlineQueryResultCachedGif)

Link to an animated GIF file stored on the Telegram servers

Tuple Fields of CachedGif

0: InlineQueryResultCachedGif

CachedMpeg4Gif(InlineQueryResultCachedMpeg4Gif)

Link to a video animation (H.264/MPEG-4 AVC video without sound) stored on the Telegram servers

Tuple Fields of CachedMpeg4Gif

0: InlineQueryResultCachedMpeg4Gif

CachedPhoto(InlineQueryResultCachedPhoto)

Link to a photo stored on the Telegram servers

Tuple Fields of CachedPhoto

0: InlineQueryResultCachedPhoto

CachedSticker(InlineQueryResultCachedSticker)

Link to a sticker stored on the Telegram servers

Tuple Fields of CachedSticker

0: InlineQueryResultCachedSticker

CachedVideo(InlineQueryResultCachedVideo)

Link to a video file stored on the Telegram servers

Tuple Fields of CachedVideo

0: InlineQueryResultCachedVideo

CachedVoice(InlineQueryResultCachedVoice)

Link to a voice message stored on the Telegram servers

Tuple Fields of CachedVoice

0: InlineQueryResultCachedVoice

Contact(InlineQueryResultContact)

Contact with a phone number

Tuple Fields of Contact

0: InlineQueryResultContact

Document(InlineQueryResultDocument)

Link to a file

Tuple Fields of Document

0: InlineQueryResultDocument

Game(InlineQueryResultGame)

Game

Tuple Fields of Game

0: InlineQueryResultGame

Gif(InlineQueryResultGif)

Link to an animated GIF file

Tuple Fields of Gif

0: InlineQueryResultGif

Location(InlineQueryResultLocation)

Location on a map

Tuple Fields of Location

0: InlineQueryResultLocation

Mpeg4Gif(InlineQueryResultMpeg4Gif)

Link to a video animation (H.264/MPEG-4 AVC video without sound)

Tuple Fields of Mpeg4Gif

0: InlineQueryResultMpeg4Gif

Photo(InlineQueryResultPhoto)

Link to a photo

Tuple Fields of Photo

0: InlineQueryResultPhoto

Venue(InlineQueryResultVenue)

Venue

Tuple Fields of Venue

0: InlineQueryResultVenue

Video(InlineQueryResultVideo)

Link to a page containing an embedded video player or a video file

Tuple Fields of Video

0: InlineQueryResultVideo

Voice(InlineQueryResultVoice)

Link to a voice recording in an .ogg container encoded with OPUS

Tuple Fields of Voice

0: InlineQueryResultVoice