tgbot::types

Enum RichBlock

pub enum RichBlock {

    Anchor(String),
    Animation(RichBlockAnimation),
    Audio(RichBlockAudio),
    BlockQuotation(RichBlockBlockQuotation),
    Buttons(RichBlockButtons),
    Collage(RichBlockCollage),
    Details(RichBlockDetails),
    Divider,
    Document(RichBlockDocument),
    ExpandableBlockQuotation(RichBlockExpandableBlockQuotation),
    Footer(RichText),
    List(Vec<RichBlockListItem>),
    Map(RichBlockMap),
    MathematicalExpression(String),
    Paragraph(RichText),
    Photo(RichBlockPhoto),
    Preformatted(RichBlockPreformatted),
    PullQuotation(RichBlockPullQuotation),
    SectionHeading(RichText, Integer),
    Slideshow(RichBlockSlideshow),
    Table(RichBlockTable),
    Thinking(RichText),
    Video(RichBlockVideo),
    VoiceNote(RichBlockVoiceNote),

}

Represents a block in a rich formatted message.

Variants

Anchor(String)

An anchor (<a> with the name attribute).

Animation(RichBlockAnimation)

An animation (<video>).

Audio(RichBlockAudio)

An audio (<audio>).

BlockQuotation(RichBlockBlockQuotation)

A block quotation (<blockquote>).

Buttons(RichBlockButtons)

A block containing a list of buttons that are show in one row, corresponding to the custom HTML tag <tg-button-row>.

Collage(RichBlockCollage)

A collage (<tg-collage>).

Details(RichBlockDetails)

An expandable block for details disclosure (<details>).

Divider

A divider (<hr />).

Document(RichBlockDocument)

A block with general file (<tg-document>).

ExpandableBlockQuotation(RichBlockExpandableBlockQuotation)

A block quotation with custom expandable attribute.

Footer(RichText)

A footer (<footer>).

List(Vec<RichBlockListItem>)

A list (<ul> or <ol>).

Map(RichBlockMap)

A map (<tg-map>).

MathematicalExpression(String)

A mathematical expression (<tg-math-block>).

Paragraph(RichText)

A paragraph (<p>).

Photo(RichBlockPhoto)

A Photo (<photo>).

Preformatted(RichBlockPreformatted)

A preformatted text (<pre> or <code>).

PullQuotation(RichBlockPullQuotation)

A quotation with centered text (<aside>).

SectionHeading(RichText, Integer)

A section heading (<h[1-6]> text, size).

Slideshow(RichBlockSlideshow)

A slideshow (<tg-slideshow>).

Table(RichBlockTable)

A table (<table>).

Thinking(RichText)

A “Thinking …” placeholder (<tg-thinking>).

The block may be used only in [crate::types::SendRichMessageDraft].

Video(RichBlockVideo)

A video (<video>).

VoiceNote(RichBlockVoiceNote)

A voice note (<audio>).