Enum tgbot::types::UpdateKind

pub enum UpdateKind {

    Message(Message),
    EditedMessage(Message),
    ChannelPost(Message),
    EditedChannelPost(Message),
    InlineQuery(InlineQuery),
    ChosenInlineResult(ChosenInlineResult),
    CallbackQuery(CallbackQuery),
    ShippingQuery(ShippingQuery),
    PreCheckoutQuery(PreCheckoutQuery),
    Poll(Poll),
    PollAnswer(PollAnswer),
    BotStatus(ChatMemberUpdated),
    UserStatus(ChatMemberUpdated),
    Unknown(JsonValue),

}

Kind of update

Variants

Message(Message)

New incoming message of any kind — text, photo, sticker, etc

Tuple Fields of Message

0: Message

EditedMessage(Message)

New version of a message that is known to the bot and was edited

Tuple Fields of EditedMessage

0: Message

ChannelPost(Message)

New incoming channel post of any kind — text, photo, sticker, etc

Tuple Fields of ChannelPost

0: Message

EditedChannelPost(Message)

New version of a channel post that is known to the bot and was edited

Tuple Fields of EditedChannelPost

0: Message

InlineQuery(InlineQuery)

New incoming inline query

Tuple Fields of InlineQuery

0: InlineQuery

ChosenInlineResult(ChosenInlineResult)

The result of an inline query that was chosen by a user and sent to their chat partner

Please see our documentation on the feedback collecting for details on how to enable these updates for your bot

Tuple Fields of ChosenInlineResult

0: ChosenInlineResult

CallbackQuery(CallbackQuery)

New incoming callback query

Tuple Fields of CallbackQuery

0: CallbackQuery

ShippingQuery(ShippingQuery)

New incoming shipping query

Only for invoices with flexible price

Tuple Fields of ShippingQuery

0: ShippingQuery

PreCheckoutQuery(PreCheckoutQuery)

New incoming pre-checkout query

Contains full information about checkout

Tuple Fields of PreCheckoutQuery

0: PreCheckoutQuery

Poll(Poll)

New poll state

Bots receive only updates about polls, which are sent or stopped by the bot

Tuple Fields of Poll

0: Poll

PollAnswer(PollAnswer)

A user changed their answer in a non-anonymous poll

Bots receive new votes only in polls that were sent by the bot itself

Tuple Fields of PollAnswer

0: PollAnswer

BotStatus(ChatMemberUpdated)

The bot’s chat member status was updated in a chat

For private chats, this update is received only when the bot is blocked or unblocked by the user

Tuple Fields of BotStatus

0: ChatMemberUpdated

UserStatus(ChatMemberUpdated)

A chat member’s status was updated in a chat

The bot must be an administrator in the chat and must explicitly specify “chat_member” in the list of allowed_updates to receive these updates.

Tuple Fields of UserStatus

0: ChatMemberUpdated

Unknown(JsonValue)

Used for unknown update types

For example, Telegram introduced a new update type, but it is not supported in old versions of tgbot

Tuple Fields of Unknown

0: JsonValue