Struct tgbot::types::ChannelChat

pub struct ChannelChat {

    pub id: ChatPeerId,
    pub title: String,
    pub accent_color: Option<AccentColor>,
    pub active_usernames: Option<Vec<String>>,
    pub available_reactions: Option<Vec<ReactionType>>,
    pub background_custom_emoji_id: Option<String>,
    pub description: Option<String>,
    pub emoji_status_custom_emoji_id: Option<String>,
    pub emoji_status_expiration_date: Option<Integer>,
    pub has_protected_content: Option<bool>,
    pub invite_link: Option<String>,
    pub linked_chat_id: Option<Integer>,
    pub message_auto_delete_time: Option<Integer>,
    pub photo: Option<ChatPhoto>,
    pub pinned_message: Option<Box<Message>>,
    pub profile_accent_color: Option<ProfileAccentColor>,
    pub profile_background_custom_emoji_id: Option<String>,
    pub username: Option<ChatUsername>,

}

Represents a channel chat.

Fields
id: ChatPeerId

Unique identifier of the channel.

title: String

Title of the channel.

accent_color: Option<AccentColor>

Identifier of the accent color for the chat name and backgrounds of the chat photo, reply header, and link preview.

Returned only in crate::types::GetChat.

active_usernames: Option<Vec<String>>

List of all active channel usernames.

Returned only in crate::types::GetChat.

available_reactions: Option<Vec<ReactionType>>

List of available reactions allowed in the chat.

If omitted, then all emoji reactions are allowed.

Returned only in crate::types::GetChat.

background_custom_emoji_id: Option<String>

Custom emoji identifier of emoji chosen by the chat for the reply header and link preview background.

Returned only in crate::types::GetChat.

description: Option<String>

Description of the channel.

Returned only in crate::types::GetChat.

emoji_status_custom_emoji_id: Option<String>

Custom emoji identifier of emoji status.

Returned only in crate::types::GetChat.

emoji_status_expiration_date: Option<Integer>

Expiration date of the emoji status of in Unix time, if any.

Returned only in crate::types::GetChat.

has_protected_content: Option<bool>

Indicates whether messages from the channel can’t be forwarded to other chats.

Returned only in crate::types::GetChat.

invite_link: Option<String>

Invite link for the channel.

Returned only in crate::types::GetChat.

linked_chat_id: Option<Integer>

Unique identifier of a linked discussion group.

Returned only in crate::types::GetChat.

message_auto_delete_time: Option<Integer>

The time after which all messages sent to the channel will be automatically deleted; in seconds.

Returned only in crate::types::GetChat.

photo: Option<ChatPhoto>

Photo associated with the channel.

Returned only in crate::types::GetChat.

pinned_message: Option<Box<Message>>

Latest pinned message in the channel.

Returned only in crate::types::GetChat.

profile_accent_color: Option<ProfileAccentColor>

Identifier of the accent color for the chat’s profile background.

Returned only in crate::types::GetChat.

profile_background_custom_emoji_id: Option<String>

Custom emoji identifier of the emoji chosen by the chat for its profile background.

Returned only in crate::types::GetChat.

username: Option<ChatUsername>

Username of the channel.

Implementations

impl ChannelChat

pub fn new<A, B>(id: A, title: B) -> Self
where A: Into<ChatPeerId>, B: Into<String>,

Creates a new ChannelChat.

Arguments

id - Unique identifier of the channel.

title - Title of the channel.

pub fn with_accent_color(self, value: AccentColor) -> Self

Sets a new accent color.

Arguments

value - Accent color for the chat.

pub fn with_active_usernames<A, B>(self, value: A) -> Self
where A: IntoIterator<Item = B>, B: Into<String>,

Sets a new list of active usernames.

Arguments

value - The list of all active channel usernames.

pub fn with_available_reactions<T>(self, value: T) -> Self
where T: IntoIterator<Item = ReactionType>,

Sets a new list of available reactions.

Arguments

value - The list of all available reactions.

pub fn with_background_custom_emoji_id<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new custom emoji identifier for the message background.

Arguments

value - Custom emoji identifier of emoji chosen by the chat for the reply header and link preview background.

pub fn with_description<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new description

Arguments

value - The description of the channel.

pub fn with_emoji_status_custom_emoji_id<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new custom emoji identifier of emoji status.

Arguments

value - Emoji identifier.

pub fn with_emoji_status_expiration_date(self, value: Integer) -> Self

Sets a new emoji status expiration date.

Arguments

value - Unix timestamp; in seconds.

pub fn with_has_protected_content(self, value: bool) -> Self

Sets a new value for a has_protected_content flag.

Arguments

value - Indicates whether messages from the chat can’t be forwarded to other chats.

pub fn with_invite_link<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new invite link.

Arguments

value - The invite link for the channel.

pub fn with_linked_chat_id(self, value: Integer) -> Self

Sets a new linked chat ID.

Arguments

value - Unique identifier of the linked discussion group.

pub fn with_message_auto_delete_time(self, value: Integer) -> Self

Sets a new message auto-delete time.

Arguments

value - Value in seconds after which all messages sent to the chat will be automatically deleted.

pub fn with_photo(self, value: ChatPhoto) -> Self

Sets a new photo.

Arguments

value - Photo associated with the channel.

pub fn with_pinned_message(self, value: Message) -> Self

Sets a new pinned message.

Arguments

value - Latest pinned message in the channel.

pub fn with_profile_accent_color(self, value: ProfileAccentColor) -> Self

Sets a new profile accent color.

Arguments

value - Accent color for the chat’s profile background.

pub fn with_profile_background_custom_emoji_id<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new custom emoji identifer for the chat’s profile background.

Arguments

value - Custom emoji identifier of the emoji chosen by the chat for its profile background.

pub fn with_username<T>(self, value: T) -> Self
where T: Into<ChatUsername>,

Sets a new username.

Arguments

value - Username of the channel.