Struct tgbot::types::InputMessageContentInvoice

pub struct InputMessageContentInvoice { /* fields omitted */ }

Represents the content of an invoice message to be sent as the result of an inline query

Implementations

impl InputMessageContentInvoice

pub fn new<A, B, C, D, E, F>(
title: A,
description: B,
payload: C,
provider_token: D,
currency: E,
prices: F
) -> Self where
A: Into<String>,
B: Into<String>,
C: Into<String>,
D: Into<String>,
E: Into<String>,
F: IntoIterator<Item = LabeledPrice>,

Creates a new InputMessageContentInvoice

Arguments

title - Product name, 1-32 characters

description - Product description, 1-255 characters

payload - Bot-defined invoice payload, 1-128 bytes. This will not be displayed to the user, use for your internal processes.

provider_token - Payment provider token, obtained via Botfather

currency - Three-letter ISO 4217 currency code, see more on currencies

prices - Price breakdown (e.g. product price, tax, discount, delivery cost, delivery tax, bonus, etc.)

pub fn max_tip_amount(self, value: Integer) -> Self

Sets the maximum accepted amount for tips in the smallest units of the currency

For example, for a maximum tip of US$ 1.45 pass max_tip_amount = 145. See the exp parameter in currencies.json, it shows the number (2 for the majority of currencies). Defaults to 0

pub fn suggested_tip_amounts(
self,
value: impl IntoIterator<Item = Integer>
) -> Self

Sets an array of suggested amounts of tip in the smallest units of the currency

At most 4 suggested tip amounts can be specified. The suggested tip amounts must be positive, passed in a strictly increased order and must not exceed max_tip_amount.

pub fn provider_data<T>(self, value: &T) -> Result<Self, JsonError> where
T: Serialize,

An object for data about the invoice, which will be shared with the payment provider. A detailed description of the required fields should be provided by the payment provider.

pub fn photo_url<T>(self, value: T) -> Self where
T: Into<String>,

Sets an URL of the product photo for the invoice

Can be a photo of the goods or a marketing image for a service. People like it better when they see what they are paying for.

pub fn photo_size(self, value: Integer) -> Self

Sets a photo size

pub fn photo_width(self, value: Integer) -> Self

Sets a photo width

pub fn photo_height(self, value: Integer) -> Self

Sets a photo height

pub fn need_name(self, value: bool) -> Self

Pass True, if you require the user’s full name to complete the order

pub fn need_phone_number(self, value: bool) -> Self

Pass True, if you require the user’s phone number to complete the order

pub fn need_email(self, value: bool) -> Self

Pass True, if you require the user’s email address to complete the order

pub fn need_shipping_address(self, value: bool) -> Self

Pass True, if you require the user’s shipping address to complete the order

pub fn send_phone_number_to_provider(self, value: bool) -> Self

Pass True, if user’s phone number should be sent to provider

pub fn send_email_to_provider(self, value: bool) -> Self

Pass True, if user’s email address should be sent to provider

pub fn is_flexible(self, value: bool) -> Self

Pass True, if the final price depends on the shipping method