tgbot::types

Struct InputRichBlock

pub struct InputRichBlock { /* private fields */ }

Represents a block in a rich formatted message to be sent.

Implementations

impl InputRichBlock

pub fn anchor<T>(name: T) -> Self
where T: Into<String>,

Creates a new InputRichBlock.

Arguments

name - The name of the anchor.

A block with an anchor, corresponding to the HTML tag <a> with the attribute name.

pub fn animation<T>(value: T) -> Self
where T: Into<InputMediaAnimation>,

Creates a new InputRichBlock.

Arguments

media - The file.

info - Additional information for the file; caption is ignored.

A block with an animation, corresponding to the HTML tag <video>.

pub fn audio<T>(value: T) -> Self
where T: Into<InputMediaAudio>,

Creates a new InputRichBlock.

Arguments

media - The file.

info - Additional information for the file; caption is ignored.

A block with a music file, corresponding to the HTML tag <audio>.

pub fn blockquote<A, B>(blocks: A, credit: Option<B>) -> Self
where A: IntoIterator<Item = InputRichBlock>, B: Into<RichText>,

Creates a new InputRichBlock.

Arguments

blocks - Content of the block.

credit - Credit of the block.

A block quotation, corresponding to the HTML tag <blockquote>.

pub fn buttons<T>(buttons: T, align: Option<RichBlockButtonsAlignment>) -> Self
where T: IntoIterator<Item = RichMessageButton>,

Creates a new InputRichBlock.

Arguments

buttons - List of buttons; 1-8.

align - Horizontal alignment.

pub fn collage<A, B>(blocks: A, caption: Option<B>) -> Self
where A: IntoIterator<Item = InputRichBlock>, B: Into<RichBlockCaption>,

Creates a new InputRichBlock.

Arguments

blocks - Elements of the collage.

caption - Caption of the block.

A collage, corresponding to the custom HTML tag <tg-collage>.

pub fn details<A, B>(summary: A, blocks: B, is_open: bool) -> Self
where A: Into<RichText>, B: IntoIterator<Item = InputRichBlock>,

Creates a new InputRichBlock.

Arguments

summary - Always shown summary of the block.

blocks - Content of the block.

is_open - Whether the content of the block is visible by default.

An expandable block for details disclosure, corresponding to the HTML tag <details>.

pub fn divider() -> Self

Creates a new InputRichBlock.

A divider, corresponding to the HTML tag <hr/>.

pub fn document<A, B>(document: A, caption: Option<B>) -> Self
where A: Into<InputMediaDocument>, B: Into<RichBlockCaption>,

Creates a new InputRichBlock.

Arguments

document - The document; caption is ignored.

caption - Caption of the block.

pub fn expandable_block_quotation<A, B>(text: A, credit: Option<B>) -> Self
where A: Into<RichText>, B: Into<RichText>,

Creates a new InputRichBlock.

Arguments

text - Content of the block.

credit - Credit of the block.

pub fn footer<T>(text: T) -> Self
where T: Into<RichText>,

Creates a new InputRichBlock.

Arguments

text - Text of the block.

A footer, corresponding to the HTML tag <footer>.

pub fn heading<T>(text: T, size: Integer) -> Self
where T: Into<RichText>,

Creates a new InputRichBlock.

Arguments

text - Text of the block.

size - Relative size of the text font; 1-6.

A section heading, corresponding to the HTML tags <h%N%>.

pub fn list<T>(items: T) -> Self
where T: IntoIterator<Item = InputRichBlockListItem>,

Creates a new InputRichBlock.

Arguments

items - Items of the list.

A list of blocks, corresponding to the HTML tag <ul> or <ol> with multiple nested tags <li>.

pub fn map<T>( location: Location, zoom: Integer, width: Integer, height: Integer, caption: Option<T>, ) -> Self
where T: Into<RichBlockCaption>,

Creates a new InputRichBlock.

Arguments

location - Location of the center of the map.

zoom - Map zoom level; 0-24.

width - Map width; 0-10000.

height - Map height; 0-10000.

caption - Caption of the block.

A block with a map, corresponding to the custom HTML tag <tg-map>. The map’s width and height must not exceed 10000 in total. The width and height ratio must be at most 20.

pub fn mathematical_expression<T>(expression: T) -> Self
where T: Into<String>,

Creates a new InputRichBlock.

Arguments

expression - The mathematical expression in LaTeX format.

A block with a mathematical expression in LaTeX format, corresponding to the custom HTML tag <tg-math-block>.

pub fn paragraph<T>(text: T) -> Self
where T: Into<RichText>,

Creates a new InputRichBlock.

Arguments

text - Text of the block.

A text paragraph, corresponding to the HTML tag <p>.

pub fn photo<T>(value: T) -> Self
where T: Into<InputMediaPhoto>,

Creates a new InputRichBlock.

Arguments

value - The file.

A block with a photo, corresponding to the HTML tag <img>.

pub fn pre<A, B>(text: A, language: Option<B>) -> Self
where A: Into<RichText>, B: Into<String>,

Creates a new InputRichBlock.

Arguments

text - Text of the block.

language - The programming language of the text.

A preformatted text block, corresponding to the nested HTML tags <pre> and <code>.

pub fn pullquote<A, B>(text: A, credit: Option<B>) -> Self
where A: Into<RichText>, B: Into<RichText>,

Creates a new InputRichBlock.

Arguments

text - Text of the block.

credit - Credit of the block.

A quotation with centered text, loosely corresponding to the HTML tag <aside>.

pub fn slideshow<A, B>(blocks: A, caption: Option<B>) -> Self
where A: IntoIterator<Item = InputRichBlock>, B: Into<RichBlockCaption>,

Creates a new InputRichBlock.

Arguments

blocks - Elements of the slideshow.

caption - Caption of the block.

A slideshow, corresponding to the custom HTML tag <tg-slideshow>.

pub fn table<T>(table: T) -> Self
where T: Into<InputRichBlockTable>,

Creates a new InputRichBlock.

Arguments

value - Content of the table.

A table, corresponding to the HTML tag <table>.

pub fn thinking<T>(text: T) -> Self
where T: Into<RichText>,

Creates a new InputRichBlock.

Arguments

text - Text of the block.

A block with a “Thinking…” placeholder, corresponding to the custom HTML tag <tg-thinking>.

The block may be used only in crate::types::SendRichMessageDraft, therefore it can’t be received in messages.

See https://t.me/addemoji/AIActions for examples of custom emoji that are recommended for usage in the block.

pub fn video<T>(value: T) -> Self
where T: Into<InputMediaVideo>,

Creates a new InputRichBlock.

Arguments

value - The file.

pub fn voice_note<T>(value: T) -> Self
where T: Into<InputMediaVoiceNote>,

Creates a new InputRichBlock.

Arguments

value - The file.