Struct tgbot::types::InvoiceParameters

pub struct InvoiceParameters {

    pub is_flexible: Option<bool>,
    pub max_tip_amount: Option<Integer>,
    pub need_email: Option<bool>,
    pub need_name: Option<bool>,
    pub need_phone_number: Option<bool>,
    pub need_shipping_address: Option<bool>,
    pub photo_height: Option<Integer>,
    pub photo_size: Option<Integer>,
    pub photo_url: Option<String>,
    pub photo_width: Option<Integer>,
    pub provider_data: Option<String>,
    pub provider_token: Option<String>,
    pub send_phone_number_to_provider: Option<bool>,
    pub send_email_to_provider: Option<bool>,
    pub suggested_tip_amounts: Option<Vec<Integer>>,

}

Represents an invoice parameters used in CreateInvoiceLink and SendInvoice.

Fields
is_flexible: Option<bool>

Indicates whether the final price depends on the shipping method.

max_tip_amount: Option<Integer>

The maximum accepted amount for tips in the smallest units of the currency.

For example, for a maximum tip of US$ 1.45 pass max_tip_amount = 145. See the exp parameter in currencies.json, it shows the number of digits past the decimal point for each currency (2 for the majority of currencies).

Defaults to 0.

need_email: Option<bool>

Indicates whether the user’s email address is required to complete the order.

need_name: Option<bool>

Indicates whether the user’s full name is required to complete the order.

need_phone_number: Option<bool>

Indicates whether the user’s phone number is required to complete the order.

need_shipping_address: Option<bool>

Indicates whether the user’s shipping address is required to complete the order.

photo_height: Option<Integer>

Photo height.

photo_size: Option<Integer>

Photo size in bytes.

photo_url: Option<String>

URL of the product photo for the invoice.

Can be a photo of the goods or a marketing image for a service.

photo_width: Option<Integer>

Photo width.

provider_data: Option<String>

Data about the invoice, which will be shared with the payment provider.

A detailed description of required fields should be provided by the payment provider.

provider_token: Option<String>

Payment provider token, obtained via @BotFather.

Pass an empty string for payments in Telegram Stars.

send_phone_number_to_provider: Option<bool>

Indicates whether the user’s phone number should be sent to the provider.

send_email_to_provider: Option<bool>

Indicates whether the user’s email address should be sent to the provider.

suggested_tip_amounts: Option<Vec<Integer>>

An array of suggested amounts of tips in the smallest units of the currency.

At most 4 suggested tip amounts can be specified. The suggested tip amounts must be positive, passed in a strictly increased order and must not exceed max_tip_amount.

Implementations

impl InvoiceParameters

pub fn with_flexible(self, value: bool) -> Self

Sets a new value for an is_flexible flag.

Arguments

value - Indicates whether the final price depends on the shipping method.

pub fn with_max_tip_amount(self, value: Integer) -> Self

Sets a new max tip amount.

Arguments

value - The maximum accepted amount for tips in the smallest units of the currency.

pub fn with_need_email(self, value: bool) -> Self

Sets a new value for a need_email flag.

Arguments

value - Indicates whether the user’s email address is required to complete the order.

pub fn with_need_name(self, value: bool) -> Self

Sets a new value for a need_name flag.

Arguments

value - Indicates whether the user’s full name is required to complete the order.

pub fn with_need_phone_number(self, value: bool) -> Self

Sets a new value for a need_phone_number flag.

Arguments

value - Indicates whether the user’s phone number is required to complete the order.

pub fn with_need_shipping_address(self, value: bool) -> Self

Sets a new value for a need_shipping_address flag.

Arguments

value - Indicates whether the user’s shipping address is required to complete the order.

pub fn with_photo_height(self, value: Integer) -> Self

Sets a new photo height.

Arguments

value - Photo height.

pub fn with_photo_size(self, value: Integer) -> Self

Sets a new photo size.

Arguments

value - Photo size in bytes.

pub fn with_photo_url<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new photo URL.

Arguments

value - Photo URL.

pub fn with_photo_width(self, value: Integer) -> Self

Sets a new photo width.

Arguments

value - Photo width.

pub fn with_provider_data<T>(self, value: &T) -> Result<Self, JsonError>
where T: Serialize,

Sets a new provider data.

Arguments

value - Data about the invoice, which will be shared with the payment provider.

pub fn with_provider_token<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new provider token.

Arguments

value - Payment provider token, obtained via @BotFather. Pass an empty string for payments in Telegram Stars.

pub fn with_send_phone_number_to_provider(self, value: bool) -> Self

Sets a new value for a send_phone_number_to_provider flag.

Arguments

value - Indicates whether the user’s phone number should be sent to the provider.

pub fn with_send_email_to_provider(self, value: bool) -> Self

Sets a new value for a send_email_to_provider flag.

Arguments

value - Indicates whether the user’s email address should be sent to the provider.

pub fn with_suggested_tip_amounts<T>(self, value: T) -> Self
where T: IntoIterator<Item = Integer>,

Sets a new list of max tip amounts.

Arguments

value - An array of suggested amounts of tips in the smallest units of the currency.