Struct tgbot::types::SendMediaGroup

pub struct SendMediaGroup { /* private fields */ }

Sends a group of photos or videos as an album.

Implementations

impl SendMediaGroup

pub fn new<T>(chat_id: T, media: MediaGroup) -> Self
where T: Into<ChatId>,

Creates a new SendMediaGroup.

chat_id - Unique identifier of the target chat.

media - Photos and videos to be sent; 2–10 items.

pub fn with_business_connection_id<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new business connection ID.

Arguments

value - Unique identifier of the business connection.

pub fn with_disable_notification(self, value: bool) -> Self

Sets a new value for a disable_notification flag.

Arguments

value - Indicates whether to send the message silently or not; a user will receive a notification without sound.

pub fn with_message_effect_id<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new message effect ID.

Arguments

value - Unique identifier of the message effect to be added to the message; for private chats only.

pub fn with_message_thread_id(self, value: Integer) -> Self

Sets a new message thread ID.

Arguments

value - Unique identifier of the target message thread; supergroups only.

pub fn with_protect_content(self, value: bool) -> Self

Sets a new value for a protect_content flag.

Arguments

value - Indicates whether to protect the contents of the sent message from forwarding and saving.

pub fn with_reply_parameters( self, value: ReplyParameters, ) -> Result<Self, ReplyParametersError>

Sets new reply parameters.

Arguments

value - Description of the message to reply to.