Struct tgbot::types::SendMessage

pub struct SendMessage { /* private fields */ }

Sends a text message.

Implementations

impl SendMessage

pub fn new<A, B>(chat_id: A, text: B) -> Self
where A: Into<ChatId>, B: Into<String>,

Creates a new SendMessage.

Arguments

chat_id - Unique identifier for the target chat.

text - Text of the message to be sent.

pub fn with_business_connection_id<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new business connection ID.

Arguments

value - Unique identifier of the business connection.

pub fn with_disable_notification(self, value: bool) -> Self

Sets a new value for a disable_notification flag.

Arguments

value - Indicates whether to send the message silently or not; a user will receive a notification without sound.

pub fn with_entities<T>(self, value: T) -> Self
where T: IntoIterator<Item = TextEntity>,

Sets a new list of entities.

Arguments

value - List of special entities that appear in the text.

Parse mode will be set to None when this method is called.

pub fn with_link_preview_options(self, value: LinkPreviewOptions) -> Self

Sets a new link preview options.

Arguments

value - Link preview generation options for the message.

pub fn with_message_effect_id<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new message effect ID.

Arguments

value - Unique identifier of the message effect to be added to the message; for private chats only.

pub fn with_message_thread_id(self, value: Integer) -> Self

Sets a new message thread ID.

Arguments

value - Unique identifier of the target message thread; supergroups only.

pub fn with_parse_mode(self, value: ParseMode) -> Self

Sets a new parse mode.

Arguments

value - Parse mode.

Entities will be set to None when this method is called.

pub fn with_protect_content(self, value: bool) -> Self

Sets a new value for a protect_content flag.

Arguments

value - Indicates whether to protect the contents of the sent message from forwarding and saving.

pub fn with_reply_markup<T>(self, value: T) -> Self
where T: Into<ReplyMarkup>,

Sets a new reply markup.

Arguments

value - Reply markup.

pub fn with_reply_parameters(self, value: ReplyParameters) -> Self

Sets new reply parameters.

Arguments

value - Description of the message to reply to.