Struct tgbot::types::Voice

pub struct Voice {
    pub duration: Integer,
    pub file_id: String,
    pub file_unique_id: String,
    pub file_size: Option<Integer>,
    pub mime_type: Option<String>,
}

Represents a voice file.

Fields
duration: Integer

Duration in seconds as defined by sender.

file_id: String

Identifier of the file.

Can be used to download or reuse the file.

file_unique_id: String

Unique identifier of the file.

It is supposed to be the same over time and for different bots. Can’t be used to download or reuse the file.

file_size: Option<Integer>

File size in bytes.

mime_type: Option<String>

MIME type as defined by sender.

Implementations

impl Voice

pub fn new<A, B>(duration: Integer, file_id: A, file_unique_id: B) -> Self
where A: Into<String>, B: Into<String>,

Creates a new Voice.

Arguments

duration - Duration in seconds.

file_id - Identifier of the file.

file_unique_id - Unique identifier of the file.

pub fn with_file_size(self, value: Integer) -> Self

Sets a new size of the file.

Arguments

value - The size of the file in bytes.

pub fn with_mime_type<T>(self, value: T) -> Self
where T: Into<String>,

Sets a new MIME type.

Arguments

value - MIME type.