tgbot::api

Trait Method

pub trait Method {
    type Response;

    // Required method
    fn into_payload(self) -> Payload;
}

Represents an API method that can be executed by the Telegram Bot API client.

Required Associated Types

type Response

The type representing a successful result in an API response.

Required Methods

fn into_payload(self) -> Payload

Converts the method into a payload for an HTTP request.