Module tgbot::methods

Methods available in the Bot API

Structs

AddStickerToSet

Add a new sticker to a set created by the bot

AnswerCallbackQuery

Send answer to callback query sent from inline keyboard

AnswerInlineQuery

Use this method to send answers to an inline query

AnswerPreCheckoutQuery

Respond to pre-checkout query

AnswerShippingQuery

Reply to shipping query

BanChatMember

Ban a user in a group, a supergroup or a channel

Close

Close the bot instance before moving it from one local server to another

CopyMessage

Copy messages of any kind

CreateChatInviteLink

Create an additional invite link for a chat

CreateNewStickerSet

Create new sticker set owned by a user

DeleteChatPhoto

Delete a chat photo

DeleteChatStickerSet

Delete a group sticker set from a supergroup

DeleteMessage

Delete a message, including service messages

DeleteMyCommands

Use this method to delete the list of the bot’s commands for the given scope and user language

DeleteStickerFromSet

Delete a sticker from a set created by the bot

DeleteWebhook

Remove webhook integration if you decide to switch back to getUpdates

EditChatInviteLink

Edit a non-primary invite link created by the bot

EditMessageCaption

Edit caption of message sent by the bot or via the bot (for inline bots)

EditMessageLiveLocation

Edit live location messages sent by the bot or via the bot (for inline bots)

EditMessageMedia

Edit audio, document, photo, or video messages

EditMessageReplyMarkup

Edit only the reply markup of messages sent by the bot or via the bot (for inline bots)

EditMessageText

Edit text and game messages sent by the bot or via the bot (for inline bots)

ExportChatInviteLink

Generate a new invite link for a chat

ForwardMessage

Forward message of any kind

GetChat

Get up to date information about the chat

GetChatAdministrators

Get a list of administrators in a chat

GetChatMember

Get information about a member of a chat

GetChatMemberCount

Get the number of members in a chat

GetFile

Get basic info about a file and prepare it for downloading

GetGameHighScores

Get data for high score tables

GetMe

Returns basic information about the bot

GetMyCommands

Use this method to get the current list of the bot’s commands

GetStickerSet

Get a sticker set

GetUpdates

Receive incoming updates using long polling

GetUserProfilePhotos

Get a list of profile pictures for a user

GetWebhookInfo

Get current webhook status

LeaveChat

Leave a group, supergroup or channel

LogOut

Log out from the cloud Bot API

PinChatMessage

Add a message to the list of pinned messages in a chat

PromoteChatMember

Promote or demote a user in a supergroup or a channel

RestrictChatMember

Restrict a user in a supergroup

RevokeChatInviteLink

Revoke an invite link created by the bot

SendAnimation

Send animation files (GIF or H.264/MPEG-4 AVC video without sound)

SendAudio

Send audio files

SendChatAction

Tell the user that something is happening on the bot’s side

SendContact

Send phone contacts

SendDice

Use this method to send a dice

SendDocument

Send general files

SendGame

Use this method to send a game

SendInvoice

Send invoice

SendLocation

Send point on the map

SendMediaGroup

Send a group of photos or videos as an album

SendMessage

Send text messages

SendPhoto

Send photo

SendPoll

Use this method to send a native poll

SendQuiz

Use this method to send a quiz

SendSticker

Send .webp sticker

SendVenue

Send information about a venue

SendVideo

Send video file

SendVideoNote

Send video message

SendVoice

Send audio files, if you want Telegram clients to display the file as a playable voice message

SetChatAdministratorCustomTitle

Set a custom title for an administrator in a supergroup promoted by the bot

SetChatDescription

Change the description of a group, a supergroup or a channel

SetChatPermissions

Set default chat permissions for all members

SetChatPhoto

Set a new profile photo for the chat

SetChatStickerSet

Set a new group sticker set for a supergroup

SetChatTitle

Change the title of a chat

SetGameScore

Set the score of the specified user in a game

SetMyCommands

Use this method to change the list of the bot’s commands

SetPassportDataErrors

Informs a user that some of the Telegram Passport elements they provided contains errors

SetStickerPositionInSet

Move a sticker in a set created by the bot to a specific position

SetStickerSetThumb

Use this method to set the thumbnail of a sticker set

SetWebhook

Specify a url and receive incoming updates via an outgoing webhook

StopMessageLiveLocation

Stop updating a live location message sent by the bot or via the bot (for inline bots) before live_period expires

StopPoll

Use this method to stop a poll which was sent by the bot

UnbanChatMember

Unban a previously kicked user in a supergroup or channel

UnpinAllChatMessages

Clear the list of pinned messages in a chat

UnpinChatMessage

Remove a message from the list of pinned messages in a chat

UploadStickerFile

Upload a .png file with a sticker for later use in createNewStickerSet and addStickerToSet methods

Traits

Method

Represents an API method

Type Definitions

StopQuiz

Use this method to stop a quiz which was sent by the bot