tgbot::types

Enum InlineQueryResult

pub enum InlineQueryResult {

    Article(InlineQueryResultArticle),
    Audio(InlineQueryResultAudio),
    CachedAudio(InlineQueryResultCachedAudio),
    CachedDocument(InlineQueryResultCachedDocument),
    CachedGif(InlineQueryResultCachedGif),
    CachedMpeg4Gif(InlineQueryResultCachedMpeg4Gif),
    CachedPhoto(InlineQueryResultCachedPhoto),
    CachedSticker(InlineQueryResultCachedSticker),
    CachedVideo(InlineQueryResultCachedVideo),
    CachedVoice(InlineQueryResultCachedVoice),
    Contact(InlineQueryResultContact),
    Document(InlineQueryResultDocument),
    Game(InlineQueryResultGame),
    Gif(InlineQueryResultGif),
    Location(InlineQueryResultLocation),
    Mpeg4Gif(InlineQueryResultMpeg4Gif),
    Photo(InlineQueryResultPhoto),
    Venue(InlineQueryResultVenue),
    Video(InlineQueryResultVideo),
    Voice(InlineQueryResultVoice),

}

Represents a result of an inline query.

Variants

Article(InlineQueryResultArticle)

Link to an article or web page.

Audio(InlineQueryResultAudio)

Link to an MP3 audio file.

CachedAudio(InlineQueryResultCachedAudio)

Link to an MP3 audio file stored on the Telegram servers.

CachedDocument(InlineQueryResultCachedDocument)

Link to a file stored on the Telegram servers.

CachedGif(InlineQueryResultCachedGif)

Link to an animated GIF file stored on the Telegram servers.

CachedMpeg4Gif(InlineQueryResultCachedMpeg4Gif)

Link to a video animation (H.264/MPEG-4 AVC video without sound) stored on the Telegram servers.

CachedPhoto(InlineQueryResultCachedPhoto)

Link to a photo stored on the Telegram servers.

CachedSticker(InlineQueryResultCachedSticker)

Link to a sticker stored on the Telegram servers.

CachedVideo(InlineQueryResultCachedVideo)

Link to a video file stored on the Telegram servers.

CachedVoice(InlineQueryResultCachedVoice)

Link to a voice message stored on the Telegram servers.

Contact(InlineQueryResultContact)

Contact with a phone number.

Document(InlineQueryResultDocument)

Link to a file.

Game(InlineQueryResultGame)

Game.

Gif(InlineQueryResultGif)

Link to an animated GIF file.

Location(InlineQueryResultLocation)

Location on a map.

Mpeg4Gif(InlineQueryResultMpeg4Gif)

Link to a video animation (H.264/MPEG-4 AVC video without sound).

Photo(InlineQueryResultPhoto)

Link to a photo.

Venue(InlineQueryResultVenue)

Venue.

Video(InlineQueryResultVideo)

Link to a page containing an embedded video player or a video file.

Voice(InlineQueryResultVoice)

Link to a voice recording in an OGG container encoded with OPUS.