Enum tgbot::types::MessageData

pub enum MessageData {

    Animation(Animation),
    Audio {
        caption: Option<Text>,
        data: Audio,
    },
    AutoDeleteTimerChanged {
        time: Integer,
    },
    ChannelChatCreated,
    ConnectedWebsite(String),
    Contact(Contact),
    DeleteChatPhoto,
    Dice(Dice),
    Document {
        caption: Option<Text>,
        data: Document,
    },
    Empty,
    Game(Game),
    GroupChatCreated,
    Invoice(Invoice),
    LeftChatMember(User),
    Location(Location),
    MigrateFromChatId(Integer),
    MigrateToChatId(Integer),
    NewChatMembers(Vec<User>),
    NewChatPhoto(Vec<PhotoSize>),
    NewChatTitle(String),
    PassportData(PassportData),
    PinnedMessage(Box<Message>),
    Photo {
        caption: Option<Text>,
        data: Vec<PhotoSize>,
    },
    Poll(Poll),
    ProximityAlertTriggered(ProximityAlertTriggered),
    Sticker(Sticker),
    SuccessfulPayment(SuccessfulPayment),
    SupergroupChatCreated,
    Text(Text),
    Venue(Venue),
    Video {
        caption: Option<Text>,
        data: Video,
    },
    VideoNote(VideoNote),
    Voice {
        caption: Option<Text>,
        data: Voice,
    },
    VoiceChatScheduled {
        start_date: Integer,
    },
    VoiceChatStarted,
    VoiceChatEnded {
        duration: Integer,
    },
    VoiceChatParticipantsInvited {
        users: Vec<User>,
    },

}

Contains message data

Variants

Animation(Animation)

Message is an animation, information about the animation

Tuple Fields of Animation

0: Animation

Audio

Audio message

Fields of Audio

caption: Option<Text>

Audio caption

data: Audio

Audio data

AutoDeleteTimerChanged

A service message about a change in auto-delete timer settings

Fields of AutoDeleteTimerChanged

time: Integer

New auto-delete time for messages in the chat

ChannelChatCreated

Service message: the channel has been created This field can‘t be received in a message coming through updates, because bot can’t be a member of a channel when it is created It can only be found in reply_to_message if someone replies to a very first message in a channel

ConnectedWebsite(String)

The domain name of the website on which the user has logged in

Tuple Fields of ConnectedWebsite

0: String

Contact(Contact)

Message is a shared contact, information about the contact

Tuple Fields of Contact

0: Contact

DeleteChatPhoto

Service message: the chat photo was deleted

Dice(Dice)

Message is a dice with random value from 1 to 6

Tuple Fields of Dice

0: Dice

Document

Document message

Fields of Document

caption: Option<Text>

Document caption

data: Document

Document data

Empty

Message has no data

Game(Game)

Message is a game, information about the game

Tuple Fields of Game

0: Game

GroupChatCreated

Service message: the group has been created

Invoice(Invoice)

Message is an invoice for a payment, information about the invoice

Tuple Fields of Invoice

0: Invoice

LeftChatMember(User)

A member was removed from the group (this member may be the bot itself)

Tuple Fields of LeftChatMember

0: User

Location(Location)

Message is a shared location, information about the location

Tuple Fields of Location

0: Location

MigrateFromChatId(Integer)

The supergroup has been migrated from a group with the specified identifier

Tuple Fields of MigrateFromChatId

0: Integer

MigrateToChatId(Integer)

The group has been migrated to a supergroup with the specified identifier

Tuple Fields of MigrateToChatId

0: Integer

NewChatMembers(Vec<User>)

New members that were added to the group or supergroup (the bot itself may be one of these members)

Tuple Fields of NewChatMembers

0: Vec<User>

NewChatPhoto(Vec<PhotoSize>)

A chat photo was change to this value

Tuple Fields of NewChatPhoto

0: Vec<PhotoSize>

NewChatTitle(String)

A chat title was changed to this value

Tuple Fields of NewChatTitle

0: String

PassportData(PassportData)

Telegram Passport data

Tuple Fields of PassportData

0: PassportData

PinnedMessage(Box<Message>)

Specified message was pinned Note that the Message object in this field will not contain further reply_to_message fields even if it is itself a reply

Tuple Fields of PinnedMessage

0: Box<Message>

Photo

Message is a photo, available sizes of the photo

Fields of Photo

caption: Option<Text>

Photo caption

data: Vec<PhotoSize>

Photos

Poll(Poll)

Message is a native poll, information about the poll

Tuple Fields of Poll

0: Poll

ProximityAlertTriggered(ProximityAlertTriggered)

Service message

A user in the chat triggered another user’s proximity alert while sharing Live Location

Tuple Fields of ProximityAlertTriggered

0: ProximityAlertTriggered

Sticker(Sticker)

Message is a sticker, information about the sticker

Tuple Fields of Sticker

0: Sticker

SuccessfulPayment(SuccessfulPayment)

Message is a service message about a successful payment, information about the payment

Tuple Fields of SuccessfulPayment

0: SuccessfulPayment

SupergroupChatCreated

Service message: the supergroup has been created This field can‘t be received in a message coming through updates, because bot can’t be a member of a supergroup when it is created It can only be found in reply_to_message if someone replies to a very first message in a directly created supergroup

Text(Text)

The actual UTF-8 text of the message, 0-4096 characters

Tuple Fields of Text

0: Text

Venue(Venue)

Message is a venue, information about the venue

Tuple Fields of Venue

0: Venue

Video

Message is a video, information about the video

Fields of Video

caption: Option<Text>

Video caption

data: Video

Video data

VideoNote(VideoNote)

Message is a video note, information about the video message

Tuple Fields of VideoNote

0: VideoNote

Voice

Message is a voice message, information about the file

Fields of Voice

caption: Option<Text>

Voice caption

data: Voice

Voice data

VoiceChatScheduled

A service message about a voice chat scheduled in the chat

Fields of VoiceChatScheduled

start_date: Integer

Point in time (Unix timestamp) when the voice chat is supposed to be started by a chat administrator

VoiceChatStarted

A service message about a voice chat started in the chat

VoiceChatEnded

A service message about a voice chat ended in the chat

Fields of VoiceChatEnded

duration: Integer

Voice chat duration; in seconds

VoiceChatParticipantsInvited

A service message about new members invited to a voice chat

Fields of VoiceChatParticipantsInvited

users: Vec<User>

New members that were invited to the voice chat